Implement ffmpeg 5.0 support
Version 5.0 of ffmpeg removed a lot of deprecated API which causes akode
to fail to build against it. The ffmpeg plugin has been amended to allow
akode to build against the new version as well as older versions, plus a
call to "av_register_all" has been re-instated for earlier versions to
ensure the player functions correctly.

Signed-off-by: aneejit1 <aneejit1@gmail.com>